Add files via upload
Welcome to a basic blog website, built using Node.js and EJS. This website is designed to be a simple yet functional platform for sharing blog posts and interacting with readers.

The homepage displays a list of all blog posts, with each post featuring a title and a brief summary. Clicking on a post takes you to a dedicated page for that post, where you can read the full article.

While this blog website is basic, it provides a great starting point for those interested in creating their own blog websites using Node.js and EJS. I want to thank Angela Yu for her tutorial on building a basic blog website using Node.js and EJS, which served as the foundation for this website.